Added support for loading DWARF line information.
The source loader doesn't try to do much. It just searches whatever
path was found in the DWARF debug information, which usually
assumes that you are on the machine and in the working directory
that was used to compile the executable. I plan to make this a
bit smarter later.